backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
pdfplumber==0.11.4
sqlmodel==0.0.22
python-dotenv==1.0.1
google-generativeai==0.8.3
